Handle the wrong datetime properly.
