+++ lib/yaml/constructor.py
@@ -11,35 +11,10 @@ __all__ = [
 from error import *
 from nodes import *
 
-import datetime
-
 import binascii, re, sys, types
 
 class ConstructorError(MarkedYAMLError):
     pass
-
-
-class timezone(datetime.tzinfo):
-    def __init__(self, offset):
-        self._offset = offset
-        seconds = abs(offset).total_seconds()
-        self._name = 'UTC%s%02d:%02d' % (
-            '-' if offset.days < 0 else '+',
-            seconds // 3600,
-            seconds % 3600 // 60
-        )
-
-    def tzname(self, dt=None):
-        return self._name
-
-    def utcoffset(self, dt=None):
-        return self._offset
-
-    def dst(self, dt=None):
-        return datetime.timedelta(0)
-
-    __repr__ = __str__ = tzname
-
 
 class BaseConstructor(object):
 
@@ -326,48 +301,9 @@ class SafeConstructor(BaseConstructor):
             raise ConstructorError(None, None,
                     "failed to decode base64 data: %s" % exc, node.start_mark)
 
-    timestamp_regexp = re.compile(
-            ur'''^(?P<year>[0-9][0-9][0-9][0-9])
-                -(?P<month>[0-9][0-9]?)
-                -(?P<day>[0-9][0-9]?)
-                (?:(?:[Tt]|[ \t]+)
-                (?P<hour>[0-9][0-9]?)
-                :(?P<minute>[0-9][0-9])
-                :(?P<second>[0-9][0-9])
-                (?:\.(?P<fraction>[0-9]*))?
-                (?:[ \t]*(?P<tz>Z|(?P<tz_sign>[-+])(?P<tz_hour>[0-9][0-9]?)
-                (?::(?P<tz_minute>[0-9][0-9]))?))?)?$''', re.X)
-
     def construct_yaml_timestamp(self, node):
         value = self.construct_scalar(node)
-        match = self.timestamp_regexp.match(node.value)
-        values = match.groupdict()
-        year = int(values['year'])
-        month = int(values['month'])
-        day = int(values['day'])
-        if not values['hour']:
-            return datetime.date(year, month, day)
-        hour = int(values['hour'])
-        minute = int(values['minute'])
-        second = int(values['second'])
-        fraction = 0
-        tzinfo = None
-        if values['fraction']:
-            fraction = values['fraction'][:6]
-            while len(fraction) < 6:
-                fraction += '0'
-            fraction = int(fraction)
-        if values['tz_sign']:
-            tz_hour = int(values['tz_hour'])
-            tz_minute = int(values['tz_minute'] or 0)
-            delta = datetime.timedelta(hours=tz_hour, minutes=tz_minute)
-            if values['tz_sign'] == '-':
-                delta = -delta
-            tzinfo = timezone(delta)
-        elif values['tz']:
-            tzinfo = timezone(datetime.timedelta(0))
-        return datetime.datetime(year, month, day, hour, minute, second, fraction,
-                                 tzinfo=tzinfo)
+        return self.generate_yaml_timestamp(value)
 
     def construct_yaml_omap(self, node):
         # Note: we do not check for duplicate keys, because it's too
