+++ tests/data/str.data
@@ -1,1 +1,5 @@
 - abcd
+- 0000-00-0
+- 2020-15-40
+- 2020-02-30
+- 2020-11-20 35:70:80
